Now the exercises are correctly generated.

+++ app/src/main/java/com/loloof64/android/basicchessendgamestrainer/PlayingActivity.kt
@@ -18,7 +18,7 @@ class PlayingActivity : AppCompatActivity(), PromotionPieceChooserDialogFragment
         val standardFEN = "rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR w KQkq - 0 1"
     }
 
-    override fun reactToPromotionPieceSelection(piece: Int) {
+    override fun reactToPromotionPieceSelection(piece: Short) {
         playingBoard.validatePromotionMove(piece)
         playingBoard.checkIfGameFinished()
         if (!playingBoard.gameFinished()) playingBoard.makeComputerPlay()
